requirements.txt
certifi==2024.8.30
fastapi==0.115.4
gradio==5.4.0
numpy==2.0.2
pandas==2.2.3
pillow==11.0.0
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
uvicorn==0.30.1